// server/enums/personalInterestsValue.ts
enum PersonalInterestsValue {
  COMMUNITY = 'COMMUNITY',
  CRAFTS = 'CRAFTS',
  CREATIVE = 'CREATIVE',
  DIGITAL = 'DIGITAL',
  KNOWLEDGE_BASED = 'KNOWLEDGE_BASED',
  MUSICAL = 'MUSICAL',
  OUTDOOR = 'OUTDOOR',
  NATURE_AND_ANIMALS = 'NATURE_AND_ANIMALS',
  SOCIAL = 'SOCIAL',
  SOLO_ACTIVITIES = 'SOLO_ACTIVITIES',
  SOLO_SPORTS = 'SOLO_SPORTS',
  TEAM_SPORTS = 'TEAM_SPORTS',
  WELLNESS = 'WELLNESS',
  OTHER = 'OTHER',
  NONE = 'NONE',
}

export default PersonalInterestsValue