// server/enums/inPrisonWorkValue.ts
enum InPrisonWorkValue {
  CLEANING_AND_HYGIENE = 'CLEANING_AND_HYGIENE',
  COMPUTERS_OR_DESK_BASED = 'COMPUTERS_OR_DESK_BASED',
  GARDENING_AND_OUTDOORS = 'GARDENING_AND_OUTDOORS',
  KITCHENS_AND_COOKING = 'KITCHENS_AND_COOKING',
  MAINTENANCE = 'MAINTENANCE',
  PRISON_LAUNDRY = 'PRISON_LAUNDRY',
  PRISON_LIBRARY = 'PRISON_LIBRARY',
  TEXTILES_AND_SEWING = 'TEXTILES_AND_SEWING',
  WELDING_AND_METALWORK = 'WELDING_AND_METALWORK',
  WOODWORK_AND_JOINERY = 'WOODWORK_AND_JOINERY',
  OTHER = 'OTHER',
}

export default InPrisonWorkValue